Project cleaned up
