--- TravelDiscovery/ContentView.swift
@@ -1,265 +1,0 @@
-//
-//  ContentView.swift
-//  TravelDiscovery
-//
-//  Created by Alejandro Robles on 02/06/23.
-//
-
-import SwiftUI
-
-struct ContentView: View {
-    init() {
-        UINavigationBar.appearance().largeTitleTextAttributes = [
-            .foregroundColor: UIColor.white
-        ]
-    }
-    var body: some View {
-        NavigationView {
-            ZStack {
-                LinearGradient(gradient:
-                                Gradient(colors:
-                                            [Color(UIColor(red: 253/255,
-                                                           green: 169/255,
-                                                           blue: 50/255,
-                                                           alpha: 1)),
-                                             Color(UIColor(red: 253/255,
-                                                           green: 130/255,
-                                                           blue: 48/255,
-                                                           alpha: 1))]),
-                               startPoint: .top,
-                               endPoint: .center)
-                    .ignoresSafeArea()
-                Color(.init(white: 0.95, alpha: 1))
-                    .offset(y: 400)
-                ScrollView {
-                    HStack {
-                        Image(systemName: "magnifyingglass")
-                        Text("Where do you want to go?")
-                        Spacer()
-                    }
-                    .font(.system(size: 14, weight: .semibold))
-                    .foregroundColor(.white)
-                    .padding()
-                    .background(Color(.init(white: 1, alpha: 0.3)))
-                    .cornerRadius(10)
-                    .padding(16)
-                    DiscoverCategoriesView()
-                    VStack {
-                        PopularDestinationsView()
-                        PopularRestaurantsView()
-                        TrendingCreatorsView()
-                    }
-                    .background(Color(.init(white: 0.95, alpha: 1)))
-                    .cornerRadius(16)
-                    .padding(.top, 32)
-                }
-                .navigationTitle("Discover")
-            }
-        }
-    }
-}
-
-struct Destination: Hashable {
-    let city, country, imageName: String
-}
-
-struct PopularDestinationsView: View {
-    
-    let destinations: [Destination] = [
-        .init(city: "Paris", country: "France", imageName: "eiffel_tower"),
-        .init(city: "Tokyo", country: "Japan", imageName: "japan"),
-        .init(city: "New York", country: "US", imageName: "new_york"),
-    ]
-
-    var body: some View {
-        VStack {
-            HStack {
-                Text("Popular Destinations")
-                    .font(.system(size: 14, weight: .semibold))
-                Spacer()
-                Text("See all")
-                    .font(.system(size: 12, weight: .semibold))
-            }
-            .padding(.horizontal)
-            .padding(.top)
-            
-            ScrollView(.horizontal) {
-                HStack(spacing: 8) {
-                    ForEach(destinations, id: \.self) { destination in
-                        
-                        VStack(alignment: .leading, spacing: 0) {
-                            Image(destination.imageName)
-                                .resizable()
-                                .scaledToFill()
-                                .frame(width: 125, height: 125)
-                                .cornerRadius(4)
-                                .padding(.all, 6)
-                            Text(destination.city)
-                                .font(.system(size: 12, weight: .semibold))
-                                .padding(.horizontal, 12)
-                            Text(destination.country)
-                                .font(.system(size: 12, weight: .semibold))
-                                .padding(.horizontal, 12)
-                                .padding(.bottom, 8)
-                                .foregroundColor(.gray)
-                        }
-                        .background(Color.white)
-                        .cornerRadius(5)
-                        .shadow(color: .init(.sRGB, white: 0.8, opacity: 1),radius: 4, x: 0, y: 2)
-                        .padding(.bottom)
-                    }
-                }
-                .padding(.horizontal)
-            }
-        }
-    }
-}
-
-struct Restaurant: Hashable {
-    let name, image: String
-}
-
-struct PopularRestaurantsView: View {
-    let restaurants: [Restaurant] = [
-        .init(name: "Japan's Finest Tapas", image: "tapas"),
-        .init(name: "Bar & Grill", image: "bar_grill")]
-    var body: some View {
-        VStack {
-            HStack {
-                Text("Popular Destinations")
-                    .font(.system(size: 14, weight: .semibold))
-                Spacer()
-                Text("See all")
-                    .font(.system(size: 12, weight: .semibold))
-            }
-            .padding(.horizontal)
-            .padding(.top)
-            ScrollView(.horizontal) {
-                HStack(spacing: 8) {
-                    ForEach(restaurants, id: \.self) { restaurant in
-                        HStack(spacing: 8) {
-                            Image(restaurant.image)
-                                .resizable()
-                                .scaledToFill()
-                                .frame(width: 60, height: 60)
-                                .clipped()
-                                .cornerRadius(5)
-                                .padding(.leading, 6)
-                                .padding(.vertical, 6)
-                            VStack(alignment: .leading, spacing: 6) {
-                                HStack {
-                                    Text(restaurant.name)
-                                    Spacer()
-                                    Button(action: {}) {
-                                        Image(systemName: "ellipsis")
-                                            .foregroundColor(.gray)
-                                    }
-                                }
-                                HStack {
-                                    Image(systemName: "star.fill")
-                                        Text("4.7 • Sushi • $$")
-                                }
-                                Text("Tokyo, Japan")
-                            }
-                            .font(.system(size: 12, weight: .semibold))
-                            Spacer()
-                        }
-                        .frame(width: 240)
-                        .background(Color.white)
-                        .cornerRadius(5)
-                        .shadow(color: .init(.sRGB, white: 0.8, opacity: 1),radius: 4, x: 0, y: 2)
-                        .padding(.bottom)
-                    }
-                }
-                .padding(.horizontal)
-            }
-        }
-    }
-}
-
-struct Users: Hashable {
-    let name, imageName: String
-}
-
-struct TrendingCreatorsView: View {
-    let users: [Users] = [
-        .init(name: "Amy Adams", imageName: "amy"),
-        .init(name: "Billy", imageName: "billy"),
-        .init(name: "Sam Smith", imageName: "sam")]
-    var body: some View {
-        VStack {
-            HStack {
-                Text("Trending Creators")
-                    .font(.system(size: 14, weight: .semibold))
-                Spacer()
-                Text("See all")
-                    .font(.system(size: 12, weight: .semibold))
-            }
-            .padding(.horizontal)
-            .padding(.top)
-            
-            ScrollView(.horizontal) {
-                HStack(alignment: .top, spacing: 12) {
-                    ForEach(users, id: \.self) { user in
-                        VStack {
-                            Image(user.imageName)
-                                .resizable()
-                                .scaledToFill()
-                                .frame(width: 60, height: 60)
-                                .cornerRadius(.infinity)
-                            Text(user.name)
-                                .font(.system(size: 11, weight: .semibold))
-                                .multilineTextAlignment(.center)
-                        }
-                        .frame(width: 60)
-                        .shadow(color: .gray, radius: 4, x: 0, y: 2)
-                        .padding(.bottom)
-                    }
-                }
-                .padding(.horizontal)
-            }
-        }
-    }
-}
-
-struct ContentView_Previews: PreviewProvider {
-    static var previews: some View {
-        ContentView()
-    }
-}
-
-struct Category: Hashable {
-    let name, imageName: String
-}
-
-struct DiscoverCategoriesView: View {
-    let categories: [Category] = [
-        .init(name: "Art", imageName: "paintpalette.fill"),
-        .init(name: "Sport", imageName: "sportscourt.fill"),
-        .init(name: "Live events", imageName: "music.mic"),
-        .init(name: "Food", imageName: "tray.fill"),
-        .init(name: "History", imageName: "books.vertical.fill"),
-    ]
-    var body: some View {
-        ScrollView(.horizontal, showsIndicators: false) {
-            HStack(alignment: .top, spacing: 14) {
-                ForEach(categories, id: \.self) { category in
-                    VStack(spacing: 8) {
-                        Image(systemName: category.imageName)
-                            .font(.system(size: 20))
-                            .foregroundColor(Color.orange)
-                            .frame(width: 64, height: 64)
-                            .background(Color.white)
-                            .cornerRadius(.infinity)
-                        Text(category.name)
-                            .font(.system(size: 12, weight: .semibold))
-                            .multilineTextAlignment(.center)
-                            .foregroundColor(.white)
-                    }
-                    .frame(width: 68)
-                }
-            }
-            .padding(.horizontal)
-        }
-    }
-}
