+++ TravelDiscovery/Discover/PopularDestinationsView.swift
@@ -1,0 +1,67 @@
+//
+//  PopularDestinationsView.swift
+//  TravelDiscovery
+//
+//  Created by Alejandro Robles on 04/06/23.
+//
+
+import SwiftUI
+
+struct PopularDestinationsView: View {
+    
+    let destinations: [Destination] = [
+        .init(city: "Paris", country: "France", imageName: "eiffel_tower"),
+        .init(city: "Tokyo", country: "Japan", imageName: "japan"),
+        .init(city: "New York", country: "US", imageName: "new_york"),
+    ]
+
+    var body: some View {
+        VStack {
+            HStack {
+                Text("Popular Destinations")
+                    .font(.system(size: 14, weight: .semibold))
+                Spacer()
+                Text("See all")
+                    .font(.system(size: 12, weight: .semibold))
+            }
+            .padding(.horizontal)
+            .padding(.top)
+            
+            ScrollView(.horizontal) {
+                HStack(spacing: 8) {
+                    ForEach(destinations, id: \.self) { destination in
+                        
+                        VStack(alignment: .leading, spacing: 0) {
+                            Image(destination.imageName)
+                                .resizable()
+                                .scaledToFill()
+                                .frame(width: 125, height: 125)
+                                .cornerRadius(4)
+                                .padding(.all, 6)
+                            Text(destination.city)
+                                .font(.system(size: 12, weight: .semibold))
+                                .padding(.horizontal, 12)
+                            Text(destination.country)
+                                .font(.system(size: 12, weight: .semibold))
+                                .padding(.horizontal, 12)
+                                .padding(.bottom, 8)
+                                .foregroundColor(.gray)
+                        }
+                        .background(Color.white)
+                        .cornerRadius(5)
+                        .shadow(color: .init(.sRGB, white: 0.8, opacity: 1),radius: 4, x: 0, y: 2)
+                        .padding(.bottom)
+                    }
+                }
+                .padding(.horizontal)
+            }
+        }
+    }
+}
+
+struct PopularDestinationsView_Previews: PreviewProvider {
+    static var previews: some View {
+        PopularDestinationsView()
+        DiscoverView()
+    }
+}
