+++ TravelDiscovery/Models/Category.swift
@@ -1,0 +1,12 @@
+//
+//  Category.swift
+//  TravelDiscovery
+//
+//  Created by Alejandro Robles on 04/06/23.
+//
+
+import Foundation
+
+struct Category: Hashable {
+    let name, imageName: String
+}
