+++ TravelDiscovery/Models/Destination.swift
@@ -1,0 +1,12 @@
+//
+//  Destination.swift
+//  TravelDiscovery
+//
+//  Created by Alejandro Robles on 04/06/23.
+//
+
+import Foundation
+
+struct Destination: Hashable {
+    let city, country, imageName: String
+}
