+++ TravelDiscovery/Models/Restaurant.swift
@@ -1,0 +1,12 @@
+//
+//  Restaurant.swift
+//  TravelDiscovery
+//
+//  Created by Alejandro Robles on 04/06/23.
+//
+
+import Foundation
+
+struct Restaurant: Hashable {
+    let name, image: String
+}
