+++ TravelDiscovery/Models/User.swift
@@ -1,0 +1,12 @@
+//
+//  User.swift
+//  TravelDiscovery
+//
+//  Created by Alejandro Robles on 04/06/23.
+//
+
+import Foundation
+
+struct Users: Hashable {
+    let name, imageName: String
+}
